requirements.txt
yfinance
pandas
requests
python-dotenv
ta
feedparser
psycopg2-binary
streamlit
vaderSentiment